:lipstick: Money.tsx UI 基本完成: Update layout styles in Money view
Applied flex properties to the TagsSection within Money.tsx to align its children flex-start and justify flex-end. Created a custom style component, MyLayout, which brings in this flex layout, and replace the use of Layout with MyLayout in the Shenzhen rendering. Also added optional className prop in Layout.tsx component.