:lipstick: Tags.tsx 完成样式 Update UI for tags and add ellipsis for long text
Modifications have been made to the UI design for tags in the application. This includes the addition of a style to handle overflows of long texts by adding an ellipsis at the end. A new AddTagButton has also been introduced at the center of the page along with slight tweaks in the global style sheet.